Improve type annotations for `io.lobster.{lobsterenv/outputs}` (#3887)
* temp save of cleaning lobsterenv

* fix type errors in lobsterenv

* temp save of lobster.outputs

* revert chemenv additional_condition rename

* first go of lobster.outputs

* fix typo in get_doc

* separate lobster inputs and outputs tests

* fix unit test

* more var name and comment cleanups

* fix my errors (some unit tests fail)

* fix some unit test

* fix unit test

* Need Confirm: change `np.float64` to float for spilling

* clarify data form in docstring

* use deprecated decorator

* pre-commit auto-fixes

* pre-commit auto-fixes

* remove DEBUG tag and var name tweaks

* add more specific types

* use PeriodicSite over Site

* standardize idx var name

* limit matplotlib version

* remove matplotlib pin

* remove left out matplotlib pin

* remove ignore override tag

* pre-commit auto-fixes

* Apply suggestions from code review


* use specific type for get_nn_info return

* add TODO for a very likely TODO comment

* remove implemented TODO tag

* replace Literal with | in docstring

* use atom{idx}_list over atom{idx}s

* revert bool description

---------
